resolved leaflet map with  window is not defined error
